Move deprecated-bare-vars tests with the rule
