+++ examples/playbooks/rule-deprecated-bare-vars-fail.yml
@@ -1,5 +1,6 @@
 ---
-- hosts: localhost
+- name: Fixture for deprecated-bare-vars rule
+  hosts: localhost
   become: false
   vars:
     my_list:
