+++ examples/playbooks/rule-deprecated-bare-vars-pass.yml
